nn_requirements.txt
matplotlib==3.7.1 
matplotlib-inline==0.1.6 
matplotlib-venn==0.11.9 
numba==0.56.4 
numpy==1.22.4 
pandas==1.5.3 
pandas-datareader==0.10.0
pandas-gbq==0.17.9
Pillow==8.4.0 
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sklearn-pandas==2.2.0 
torch @ https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp310-cp310-linux_x86_64.whl#sha256=a7a49d459bf4862f64f7bc1a68beccf8881c2fa9f3e0569608e16ba6f85ebf7b 
torch-geometric==2.3.1 
torch-scatter==2.1.1 
torchaudio @ https://download.pytorch.org/whl/cu118/torchaudio-2.0.2%2Bcu118-cp310-cp310-linux_x86_64.whl#sha256=26692645ea061a005c57ec581a2d0425210ac6ba9f923edf11cc9b0ef3a111e9
torchdata==0.6.1 
torchmetrics==0.11.4 
torchsummary==1.5.1 
torchtext==0.15.2 
torchvision @ https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp310-cp310-linux_x86_64.whl#sha256=19ca4ab5d6179bbe53cff79df1a855ee6533c2861ddc7389f68349d8b9f8302a 
tqdm==4.65.0 
